requirements.txt
discord.py>=2.4.0
streamlit>=1.37.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
pydantic>=2.8.0
